refatorar e unificar availabilities
