+++ src/components/AvailabilityFriendProfile/index.tsx
@@ -1,10 +1,11 @@
-import { Card } from "react-bootstrap";
+import { useState } from "react";
+import { Card, ToggleButton } from "react-bootstrap";
 import { useQuery } from "react-query";
 import api from "../../services/api";
-import { HoursButton } from "../HoursButton";
+import { formatHour } from "../../utils/formatHour";
 import styles from './styles.module.scss';
 
-type AvailabilityFriendProfile = {
+export type AvailabilityProfile = {
     user: {
         name: string,
         id: number,
@@ -13,12 +14,14 @@ type AvailabilityFriendProfile = {
     }
 }
 
-type Profile = {
+export type Profile = {
     ProfileAvailability: [
         {
             availability: {
+                id: number,
                 day: Date,
-                hour: Date
+                start_time: Date,
+                end_time: Date
             }
         }
     ]
@@ -29,19 +32,39 @@ type Profile = {
 //trazer os resultados de horários disponíveis
 
 export function AvailabilityFriendProfile() {
-    const { data } = useQuery<AvailabilityFriendProfile>('availabilities', async () => {
+    const { data } = useQuery<AvailabilityProfile>('availabilitiesFriends', async () => {
         const response = await api.get('/user/availability/1')
         return response.data;
     }, {
         staleTime: 1000 * 60, //cache 1 minute
     })
 
+    const [radioValue, setRadioValue] = useState(0);
+
     return (
         <Card style={{ border: '0' }}>
             <Card.Body className={styles.availabilityContainer}>
                 <Card.Title style={{ fontWeight: '700', fontSize: '1rem' }}>Escolha o melhor horário</Card.Title>
                 <Card.Body className={styles.availabilityCardContainer}>
-                    <HoursButton />
+                    {data?.user.Profile.ProfileAvailability.map((item, idx) => {
+                        let startTime = formatHour(item.availability.start_time);
+                        let endTime = formatHour(item.availability.end_time);
+
+                        return (
+                            <ToggleButton
+                                key={idx}
+                                id={`radio-${idx}`}
+                                type="radio"
+                                variant={'outline-dark'}
+                                checked={radioValue === item.availability.id}
+                                value={item.availability.id}
+                                onChange={(e) => setRadioValue(Number(e.currentTarget.value))}
+                                className={styles.availabilityButton}
+                            >
+                                {`${startTime} - ${endTime}`}
+                            </ToggleButton>
+                        )
+                    })}
                 </Card.Body>
             </Card.Body>
         </Card>
