+++ src/components/AvailabilityLoggedUser/styles.module.scss
@@ -16,3 +16,8 @@
     text-align: center;
     color: var(--gray-100);
 }
+
+.badge {
+    font-size: 1rem;
+    padding: 0.5rem;
+}
