--- src/components/HoursButton/styles.module.scss
@@ -1,10 +1,0 @@
-.availabilityButton {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background: var(--secondary-07);
-    border-radius: 1rem;
-    font-size: 1rem;
-    color: var(--black);
-    padding: 0.5rem;
-}
