--- src/components/HoursList/styles.module.scss
@@ -1,4 +1,0 @@
-.badge {
-    font-size: 1rem;
-    padding: 0.5rem;
-}
